solving warnings in Analysis.cpp and IO/OutputHandler.cpp

+++ Analysis.cpp
@@ -313,7 +313,7 @@ int Analysis::RunAnalysis(const fs::path& filename, const fs::path& path)
                exit(EXIT_FAILURE);
           }
 #pragma omp parallel for
-          for(signed int i=0; i<_areaForMethod_A.size(); i++)
+          for(long unsigned int i=0; i < _areaForMethod_A.size(); i++)
           {
                Method_A method_A ;
                method_A.SetMeasurementArea(_areaForMethod_A[i]);
@@ -340,7 +340,7 @@ int Analysis::RunAnalysis(const fs::path& filename, const fs::path& path)
           }
 
 #pragma omp parallel for
-          for(signed int i=0; i<_areaForMethod_B.size(); i++)
+          for(long unsigned int i=0; i < _areaForMethod_B.size(); i++)
           {
                Method_B method_B;
                method_B.SetMeasurementArea(_areaForMethod_B[i]);
@@ -364,7 +364,7 @@ int Analysis::RunAnalysis(const fs::path& filename, const fs::path& path)
                exit(EXIT_FAILURE);
           }
 #pragma omp parallel for
-          for(signed int i=0; i<_areaForMethod_C.size(); i++)
+          for(long unsigned int i=0; i < _areaForMethod_C.size(); i++)
           {
                Method_C method_C;
                method_C.SetMeasurementArea(_areaForMethod_C[i]);
@@ -398,7 +398,7 @@ int Analysis::RunAnalysis(const fs::path& filename, const fs::path& path)
           }
 
 #pragma omp parallel for
-          for(signed int i=0; i<_areaForMethod_D.size(); i++)
+          for(long unsigned int i=0; i<_areaForMethod_D.size(); i++)
           {
                Method_D method_D;
                method_D.SetStartFrame(_StartFramesMethodD[i]);
